Move CLI parsing to common module
