+++ components/src/dynamo/common/utils/__init__.py
@@ -8,10 +8,11 @@ This module contains shared utility functions used across multiple
 Dynamo backends and components.
 
 Submodules:
+    - endpoint_types: Endpoint type parsing utilities
     - paths: Workspace directory detection and path utilities
     - prometheus: Prometheus metrics collection and logging utilities
 """
 
-from dynamo.common.utils import paths, prometheus
+from dynamo.common.utils import endpoint_types, paths, prometheus
 
-__all__ = ["paths", "prometheus"]
+__all__ = ["endpoint_types", "paths", "prometheus"]
