+++ components/src/dynamo/trtllm/main.py
@@ -37,6 +37,7 @@ from transformers import AutoConfig
 
 import dynamo.nixl_connect as nixl_connect
 from dynamo.common.config_dump import dump_config
+from dynamo.common.utils.endpoint_types import parse_endpoint_types
 from dynamo.common.utils.prometheus import register_engine_metrics_callback
 from dynamo.llm import ModelInput, ModelRuntimeConfig, ModelType, register_llm
 from dynamo.runtime import DistributedRuntime
@@ -61,26 +62,6 @@ from dynamo.trtllm.utils.trtllm_utils import (
 DEFAULT_KV_EVENT_BUFFER_MAX_SIZE = 1024
 
 configure_dynamo_logging()
-
-
-def parse_endpoint_types(endpoint_types_str: str) -> ModelType:
-    """Parse endpoint types string into ModelType flags."""
-    types = [t.strip().lower() for t in endpoint_types_str.split(",")]
-
-    result = None
-    for t in types:
-        if t == "chat":
-            flag = ModelType.Chat
-        elif t == "completions":
-            flag = ModelType.Completions
-        else:
-            raise ValueError(
-                f"Invalid endpoint type: '{t}'. Valid options: 'chat', 'completions'"
-            )
-
-        result = flag if result is None else result | flag
-
-    return result
 
 
 async def graceful_shutdown(runtime):
