+++ components/src/dynamo/vllm/main.py
@@ -15,6 +15,7 @@ from vllm.v1.engine.async_llm import AsyncLLM
 from vllm.v1.metrics.prometheus import setup_multiprocess_prometheus
 
 from dynamo.common.config_dump import dump_config
+from dynamo.common.utils.endpoint_types import parse_endpoint_types
 from dynamo.llm import (
     ModelInput,
     ModelRuntimeConfig,
@@ -239,26 +240,6 @@ def setup_vllm_engine(config, stat_logger=None):
     logger.info(f"VllmWorker for {config.served_model_name} has been initialized")
 
     return engine_client, vllm_config, default_sampling_params, prometheus_temp_dir
-
-
-def parse_endpoint_types(endpoint_types_str: str) -> ModelType:
-    """Parse endpoint types string into ModelType flags."""
-    types = [t.strip().lower() for t in endpoint_types_str.split(",")]
-
-    result = None
-    for t in types:
-        if t == "chat":
-            flag = ModelType.Chat
-        elif t == "completions":
-            flag = ModelType.Completions
-        else:
-            raise ValueError(
-                f"Invalid endpoint type: '{t}'. Valid options: 'chat', 'completions'"
-            )
-
-        result = flag if result is None else result | flag
-
-    return result
 
 
 async def register_vllm_model(
